OData V2 Edm.Binary
