+++ lib/csdl2openapi.js
@@ -2479,8 +2479,14 @@ module.exports.csdl2openapi = function (
         break;
       case "Edm.Binary":
         s = oas31
-          ? { type: "string", contentEncoding: "base64url" }
-          : { type: "string", format: "base64url" };
+          ? {
+              type: "string",
+              contentEncoding: csdl.$Version < "4.0" ? "base64" : "base64url",
+            }
+          : {
+              type: "string",
+              format: csdl.$Version < "4.0" ? "byte" : "base64url",
+            };
         if (element.$MaxLength)
           s.maxLength = Math.ceil((4 * element.$MaxLength) / 3);
         break;
